Add networking support
