+++ Assets/Game/MainSettings.lighting
