+++ Assets/Game/MainSettings.lighting.meta
