+++ Assets/Game/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4ca03010f6393172b9bb8c2c27ff3a2a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
