+++ Assets/Game/Prefabs/Player.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: a5eed22e3988d729e808c16f602b3974
-folderAsset: yes
-DefaultImporter:
+guid: 22b6fc01ed2c23789bcc7aaf1683d711
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
