+++ Assets/NetworkSetup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9bdb3c46d695ed94794dd6709199c765
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
