+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,6 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/Main.unity
+    path: Assets/Game/Main.unity
     guid: 2cda990e2423bbf4892e6590ba056729
   m_configObjects: {}
